migrate to ts: src/handler/cal_event_handler.ts
